% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaggedFun.R
\name{BaggedFun}
\alias{BaggedFun}
\title{To evaluate the instability of the curve}
\usage{
BaggedFun(data, formula, B, select = 0.05, alpha = 0.05,
  family = gaussian)
}
\arguments{
\item{data}{data set}

\item{formula, }{get from as.formula()}

\item{B}{bootstrap times}

\item{select, alpha}{same as mpf funciton,default=0.05}

\item{family}{default as gaussian, alternatively, cox}
}
\value{
a list contains the bagged fitted values, variance, conditional variance, and deviation.
}
\description{
This function return a list contains the bagged fitted values, variance, conditional variance, and deviation.
}
\examples{
library(mfp)
formula<-as.formula(y ~ fp(x1,df=4) + fp(x3,df=4) + fp(x5,df=4) + fp(x6,df=4) + fp(x7,df=4) + fp(x10,df=4)
+x9+ x2 +x4a + x4b + x8)
BaggedFun(data=data,B=10,formula=formula)

}
\seealso{
mfp() Patrial()
}
