% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEOI.R
\name{DataEOI}
\alias{DataEOI}
\title{To obtain hat values, studenlized residual, and Cook distance}
\usage{
DataEOI(data, SampleName = "")
}
\arguments{
\item{data}{results from lm() or mfp()}

\item{SampleName}{which variable in the original data is considered to be sample labels, defaulted as the row name of the data}
}
\value{
A dataframe
}
\description{
This function returns a data.frame contains hat values, Studentized Residuals,  Cook distance, and sample number
}
\examples{

library(mfp)
library(PlotForMFP)
re<-mfp(y ~ fp(x1,df=4) + fp(x3,df=4) + fp(x5,df=4) + fp(x6,df=4) + fp(x7,df=4) + fp(x10,df=4)
        + x9  + x2 +x4a  + x4b + x8, data=DataMFP,select = 0.05,rescale = FALSE)
View(DataEOI(re))
PlotEOI(re)


}
