% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSpline.R
\name{FitSpline}
\alias{FitSpline}
\title{Fiited the splines.}
\usage{
FitSpline(data, x, y, alpha = 0.05, cv = TRUE, span = 0.75,
  B = 100, m = 200, method = "loess")
}
\arguments{
\item{data}{input data,default as a data.frame}

\item{x, y}{variable and reponse}

\item{alpha}{range of confidence intervals, default as 0.05}

\item{span}{special for loess method, values to be used.}

\item{B}{times of bootstrap, default as 100}

\item{m}{special for loess method, the number of predicted values to draw the plot}

\item{method}{default as loess, weighted and bootstrap are options.}
}
\value{
A list contains initial data, fitted values, 95% confidence intervals
}
\description{
This function fits the spline using three methods, namely, weighted,  boostrap, and loess. Actually, the first two methods are the same
cubic splines. The difference is that weighted method estimates confidence interval using fuctional form, while the boostrap method estimates
confidence interval using bootstrap method. FitSpline returns a list contains initial data, fitted values, and confidence intervals, which can
be applied to make spline plot, see PlotSpline().
}
\examples{
x<-rnorm(100,5,2)
y=x^2+rnorm(100,0,5)-50
data<-data.frame(x=x,y=y)
temp<-FitSpline(data,x="x",y="y")
PlotSpline(temp,col.CI=rgb(70,130,180,maxColorValue = 255),xlab="X",ylab="y")
}
\seealso{
PlotSpline
}
