% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate Data.R
\name{GenData}
\alias{GenData}
\title{To generate simulated data}
\usage{
GenData(n, rmatrix, mu = 0, se = 1)
}
\arguments{
\item{n}{the sample size}

\item{rmatrix}{correlation matrix, obtained from GenMatrix}

\item{mu, }{mean of the variabels, default as 0, also can can set as vector,e.g. c(1,2,3,4)}

\item{se, }{se of the variabels, default as 1, also can can set as vector,e.g. c(1,2,3,4)}
}
\value{
a matrix
}
\description{
This function generate random simulated data, with certain correlation between variables,
}
\examples{

rlist=list(c(1,5,0.7),c(1,10,0.5),c(2,6,0.5),c(4,8,-0.7),c(7,8,0.3),c(11,12,0.7),c(9,13,0.5),c(7,14,0.5))
rmatrix<-GenMatrix(15,rlist)
D<-GenData(100,rmatrix,mu=0,se=1.25)
}
\seealso{
GenMatrix
}
