% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate Data.R
\name{GenMatrix}
\alias{GenMatrix}
\title{To generate matrix, can be applied to GenData}
\usage{
GenMatrix(n, rlist = list(), a = 0, b = 1)
}
\arguments{
\item{n}{the dimension of matrix}

\item{rlist}{a list to add values to matrix}

\item{a, }{default values of the matrix, defaulted as 0}

\item{b, }{default values of the diag of the matrix, defaulted as 1}
}
\value{
A matrix
}
\description{
This function generate matrix, expecially correlation matrix, which can be applied to GenData function
}
\examples{

rlist=list(c(1,5,0.7),c(1,10,0.5),c(2,6,0.5),c(4,8,-0.7),c(7,8,0.3),c(11,12,0.7),c(9,13,0.5),c(7,14,0.5))
rmatrix<-GenMatrix(15,rlist)
D<-GenData(100,rmatrix,mu=0,se=1.25)
}
\seealso{
GenData
}
