% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEOI.R
\name{PlotEOI}
\alias{PlotEOI}
\title{To get a plot to reflect the hat values, Studentized Residuals,  Cook distance}
\usage{
PlotEOI(data, SampleName = "", ylim = c(), xlim = c(),
  CI.border = NA, CI.color = "grey", CutOff.Lev = "",
  Lev.lty = "dashed", Lev.col = "black", Lev.cex = 1,
  CutOff.RS = 3, RS.lty = "dashed", RS.col = "black", RS.cex = 1,
  CutOff.CookD = "", CookD.col = "red")
}
\arguments{
\item{data}{results from lm() or mfp()}

\item{SampleName}{which variable in the original data is considered to be sample labels, defaulted as the row name of the data}
}
\value{
A plot
}
\description{
EOI is short for extreme values, outliers, and influencial values. This function returns a nice plot contains these information
}
\examples{

library(mfp)
library(PlotForMFP)
re<-mfp(y ~ fp(x1,df=4) + fp(x3,df=4) + fp(x5,df=4) + fp(x6,df=4) + fp(x7,df=4) + fp(x10,df=4)
        + x9  + x2 +x4a  + x4b + x8, data=DataMFP,select = 0.05,rescale = FALSE)
View(DataEOI(re))
PlotEOI(re)


}
