% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPatrial.R
\name{PlotPatrial}
\alias{PlotPatrial}
\title{To make patrial plot for MFP}
\usage{
PlotPatrial(data, var = "", xlim = data$xlim, ylim = data$ylim,
  xlab = data$xlab, ylab = data$ylab, log = data$log,
  shadow = FALSE, shadow.col = rgb(215, 215, 215, maxColorValue = 255),
  point = TRUE, p.pch = 1, p.col = "black", p.cex = 1,
  l.lty = "solid", l.col = "black", l.cex = 1, CI.border = NA,
  CI.col = "grey")
}
\arguments{
\item{data}{results obtained from the Patrial()}

\item{xlim, ylim}{a list, defaulted as the range of x and y. to set them, using ylim=list(var1=c(1,10),...), where var1 is the variable in the final function}

\item{xlab, ylab}{a list, the labels of axis. To set it, using xlab=list(var1="name1",var2="name2",...)}

\item{log}{a list, which axis needs the log scale, defaulted as "", if "x", the x axis scaled; if "xy", both scaled. To set it, using log=list(var1="xy",var2="x",...)}

\item{shadow}{logical, whether is 95% sample areas should be maked as a shadowed rectangle}

\item{shadow.col}{the color of the shadow, defaulted as rgb(215,215,215,maxColorValue = 255)}

\item{p.pch}{character of the points, defaulted as 1}

\item{p.col}{color of the points, defaulted as "black"}

\item{p.cex}{size of the points, defaulted as 1}

\item{l.lty}{type of the line, defaulted as "solid"}

\item{l.col}{color of the line, defaulted as "black"}

\item{l.cex}{size of the line, defaulted as 1}

\item{CI.border}{logical, whether the border should be drawn for the CI, default as NA}

\item{CI.col}{the color for the CI band, defaulted as "grey"}
}
\value{
A patrial plot with confidence interval
}
\description{
This function returns with a nice patrial plot with confidence interval using the results from mfp function or Patrial()
}
\examples{

x<-runif(100,0,10)
y<-runif(100,0,10)
z=log(x)+0.2*y^2 + rnorm(100,0,2)

library(mfp)
re<-mfp(z~fp(x,4)+fp(y,4))
re0<-Patrial("y",re)
PlotPatrial(re0)
PlotPatrial(re,xlab=list(x="X in R"))
PlotPatrial(re,"x")

}
\seealso{
PlotPatrial(),mfp package
}
