% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotResiduals.R
\name{PlotResiduals}
\alias{PlotResiduals}
\title{To Plot residuals for MPF model or general linear model}
\usage{
PlotResiduals(data, x, alpha = 0.05, cv = TRUE, span = 0.75,
  B = 100, m = 200, method = "loess", cex = 1, xlab = "",
  ylab = "", log = "", points = FALSE, col.CI = "royalblue4",
  xlim = "", ylim = "", col.shadow = rgb(215, 215, 215, maxColorValue
  = 255), shadow = TRUE, tick.length = 0.03, tick = FALSE)
}
\arguments{
\item{x}{the variable to make residual plot}

\item{xlab, ylab}{the label of axis.}

\item{log}{which axis need a log scale, e.g. "x","y","xy"}

\item{points}{logical, should be points be plotted, default as FALSE}

\item{col.CI}{colour of confidence interval,default as "royalblue4"}

\item{xlim, ylim}{limits of axis, default as the range of y}

\item{col.shadow}{color of the shadow}

\item{shadow}{logical, should be shadow be plotted?, default as TRUE}

\item{tick.length}{the length of ticks,defaults as 0.03.}

\item{tick}{logical, should be ticks be plotted? default as TRUE}

\item{Data}{data from mfp() or lm(), or FitSpline()}
}
\value{
A smoothed plot with confidence interval
}
\description{
This function returns a nice residual plot with confidence interval. The only input data could be results from mfp() or lm(), and a variable.
}
\examples{
x<-rnorm(100,5,2)
y=x^2+rnorm(100,0,5)-50
data<-data.frame(x=x,y=y)
temp<-FitModel(data,x="x",y="y")
PlotSpline(temp,col.CI=rgb(70,130,180,maxColorValue = 255),xlab="X",ylab="y")
}
\seealso{
FitSpline(),PlotSpline()
}
