% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpline.R
\name{PlotSpline}
\alias{PlotSpline}
\title{To make a spline plot}
\usage{
PlotSpline(data, x, y, alpha = 0.05, cv = TRUE, span = 0.75,
  B = 100, m = 200, method = "loess", cex = 1, xlab = "",
  ylab = "", log = "", points = TRUE, col.CI = "royalblue4",
  xlim = "", ylim = "", col.shadow = rgb(215, 215, 215, maxColorValue
  = 255), shadow = FALSE, tick.length = 0.03, tick = TRUE)
}
\arguments{
\item{xlab, ylab}{the label of axis.}

\item{log}{which axis need a log scale, e.g. "x","y","xy"}

\item{points}{logical, should be points be plotted, default as FALSE}

\item{col.CI}{colour of confidence interval,default as "royalblue4"}

\item{xlim, ylim}{limits of axis, default as the range of y}

\item{col.shadow}{color of the shadow}

\item{shadow}{logical, should be shadow be plotted?, default as TRUE}

\item{tick.length}{the length of ticks,defaults as 0.03.}

\item{tick}{logical, should be ticks be plotted? default as TRUE}

\item{SpList}{list from FiitModel}
}
\value{
A smoothed plot with confidence interval
}
\description{
This function returns a spline plot with confidence interval. This function is based on the basic plot, so a lot of parameters
can be changed. Be aware that not all elements are needed. The priciple of graphic is as simple as possible.
}
\examples{
x<-rnorm(100,5,2)
y=x^2+rnorm(100,0,5)-50
data<-data.frame(x=x,y=y)
temp<-FitModel(data,x="x",y="y")
PlotSpline(temp,col.CI=rgb(70,130,180,maxColorValue = 255),xlab="X",ylab="y")
}
