% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate Data.R
\name{resampler}
\alias{resampler}
\title{To generate bootstap sample with equal sample size}
\usage{
resampler(data)
}
\arguments{
\item{data}{original data}
}
\value{
a bootstrap sample
}
\description{
This function generate bootstap sample with equal sample size
}
\examples{
library(PlotForMFP)
resampler(DataMFP)

}
